+++ src/modules/config.ts
@@ -5,8 +5,10 @@ namespace Config {
     export const nicknameMapping = Secret.nicknameMapping;
     export const tempMasseurData = Secret.tempMasseurData;
     export const EXCLUDENAMES = Secret.EXCLUDENAMES;
+    export const DateRange = "B1:O1"
     export const flags = {
         ignorePastDays: true,
         InviteEveryoneByEmailPrefix: false,
+        skipAfterThursday: true,
     }
 }